groundstation: Move most of implementation to `internal/groundstation`.
