+++ internal/groundstation/gpustats.go
@@ -1,4 +1,4 @@
-package remote
+package groundstation
 
 import (
 	"log/slog"
@@ -7,7 +7,7 @@ import (
 	"github.com/gpuctl/gpuctl/internal/uplink"
 )
 
-func HandleStatusSubmission(stats uplink.GPUStats, req *http.Request, log *slog.Logger) error {
+func (gs *groundstation) gpustats(stats uplink.GPUStats, req *http.Request, log *slog.Logger) error {
 	log.Info("Got GPU stats", "stats", stats)
 	return nil
 }
